* embassyApi refactoring
