+++ package.json
@@ -7,9 +7,9 @@
     "start": "node bot.js",
     "dev": "nodemon bot.js",
     "embassy-tunnel": "ssh -f -NR 8001:127.0.0.1:8001 root@nickkiselev.me -o ServerAliveInterval=30 -o ServerAliveCountMax=3 -o ExitOnForwardFailure=yes",
-    "embassy-api": "npm run embassy-tunnel && nohup node ./services/embassyApi.js &",
-    "embassy-api-service": "/usr/bin/npm run embassy-tunnel && /usr/bin/node ./services/embassyApi.js",
-    "stop-embassy-api": "skill -c 'nohup node ./services/embassyApi.js &'",
+    "embassy-api": "npm run embassy-tunnel && nohup node embassyApi.js &",
+    "embassy-api-service": "/usr/bin/npm run embassy-tunnel && /usr/bin/node embassyApi.js",
+    "stop-embassy-api": "skill -c 'nohup node embassyApi.js &'",
     "git": "git add * && git commit && git push",
     "git-redo": "nodemon bot.js",
     "update-and-start": "npm run kill && git pull && npm i && npm run start-bot && npm run start-api-https",
